+++ src/Molecules/AccordionItem/AccordionItem.types.ts
@@ -20,3 +20,15 @@ export type Props = {
   pl?: number;
   pr?: number;
 };
+
+export type ItemProps = {
+  hasFocus: boolean;
+  disableBackgroundColor?: boolean;
+  p?: number;
+  px?: number;
+  py?: number;
+  pt?: number;
+  pb?: number;
+  pl?: number;
+  pr?: number;
+};