+++ src/Molecules/Pagination/Pagination.types.ts
@@ -1,21 +1,35 @@
-export type PaginationProps = {
+type PartialKeys<T, K extends keyof T> = Partial<Pick<T, K>> & Omit<T, K>;
+
+type PaginationPropsBase = {
   variant: 'regular' | 'large' | 'compact';
   currentPage?: number;
   itemsPerPage: number;
   totalItems: number;
-  onPageChange: (newPage: number) => void;
+  onPageChange?: (newPage: number) => void;
   label?: string;
   nextPageLabel?: string;
   previousPageLabel?: string;
   currentPageLabel?: string;
   pageItemLabel?: string;
+  getPageHref?: (pageNumber: number) => string;
 };
+
+interface WithGetPageHref {
+  getPageHref: NonNullable<PaginationPropsBase['getPageHref']>;
+}
+
+interface WithOnPageChange {
+  onPageChange: NonNullable<PaginationPropsBase['onPageChange']>;
+}
+
+export type PaginationProps = PaginationPropsBase & (WithGetPageHref | WithOnPageChange);
 
 export type PageItemProps = {
   isCurrentPage?: boolean;
   onClick?: () => void;
   children: React.ReactText;
   label: string;
+  href?: string;
 };
 
 export type PageItemsProps = {
@@ -26,12 +40,17 @@ export type PageItemsProps = {
   TruncatedPageNumbers: React.FC;
   currentPageLabel: string;
   pageItemLabel: string;
+  getPageHref?: PaginationPropsBase['getPageHref'];
 };
 
 export type BrowseButtonProps = {
   direction: 'left' | 'right';
   onClick: () => void;
   label: string;
+};
+
+export type BrowseLinkProps = PartialKeys<BrowseButtonProps, 'onClick'> & {
+  href: NonNullable<PageItemProps['href']>;
 };
 
 export type PaginationCompactProps = {
@@ -48,3 +67,15 @@ export type PaginationDefaultProps = PaginationCompactProps & {
   currentPageLabel: string;
   pageItemLabel: string;
 };
+
+export type PaginationCompactLinkProps = PartialKeys<
+  PaginationCompactProps,
+  'onClickPrevious' | 'onClickNext'
+> &
+  WithGetPageHref;
+
+export type PaginationDefaultLinkProps = PartialKeys<
+  PaginationDefaultProps,
+  'onClickPrevious' | 'onClickNext' | 'onClickPageItem'
+> &
+  WithGetPageHref;